final project but with output loaded
The code for this file is the same as the file uploaded before. I realized that it might be more helpful to upload a file that has the output from the epochs loaded. So in this file, we would be able to see the Ukiyo-e-style images generated by the generator against the pictures it has applied the transformations to. There are no other changes.